feat: create/update tests for events + add javadocs to events and listeners
