+++ src/main/java/edu/ntnu/idi/idatt/domain/event/EventBus.java
@@ -8,26 +8,26 @@ package edu.ntnu.idi.idatt.domain.event;
  * @author Gilianne Reyes
  */
 public interface EventBus {
-    /**
-     * Registers an event handler.
-     *
-     * @param eventType is the type of event to register the handler for.
-     * @param handler is the event handler to register.
-     */
-    <E extends GameEvent> void register(Class<E> eventType, EventHandler<E> handler);
+  /**
+   * Registers an event handler.
+   *
+   * @param eventType is the type of event to register the handler for.
+   * @param handler is the event handler to register.
+   */
+  <E extends GameEvent> void register(Class<E> eventType, EventHandler<E> handler);
 
-    /**
-     * Unregisters an event handler.
-     *
-     * @param eventType is the type of event to unregister the handler for.
-     * @param handler is the event handler to unregister.
-     */
-    <E extends GameEvent> void unregister(Class<E> eventType, EventHandler<E> handler);
+  /**
+   * Unregisters an event handler.
+   *
+   * @param eventType is the type of event to unregister the handler for.
+   * @param handler is the event handler to unregister.
+   */
+  <E extends GameEvent> void unregister(Class<E> eventType, EventHandler<E> handler);
 
-    /**
-     * Publishes an event to all registered event handlers.
-     *
-     * @param event is the event to publish.
-     */
-    void publish(GameEvent event);
+  /**
+   * Publishes an event to all registered event handlers.
+   *
+   * @param event is the event to publish.
+   */
+  void publish(GameEvent event);
 }
