+++ src/main/java/edu/ntnu/idi/idatt/domain/event/GameEvent.java
@@ -1,11 +1,14 @@
 package edu.ntnu.idi.idatt.domain.event;
 
 /**
- * Interface for game events.
+ * Marker interface for all game events in the system.
+ *
+ * <p>Types implementing this interface represent discrete events
+ * that occur during gameplay (e.g. player moved, turn started).</p>
  *
  * @version 0.1
  * @since 0.1
  * @author Gilianne Reyes
  */
-public interface GameEvent {
-}
+public interface GameEvent {}
+
