+++ src/main/java/edu/ntnu/idi/idatt/domain/event/common/PlayerMovedEvent.java
@@ -7,8 +7,14 @@ import edu.ntnu.idi.idatt.domain.event.GameEvent;
 /**
  * Represents an event where a player moves from one tile to another.
  *
- * @author Gilianne Reyes
+ * <p>Encapsulates the {@link Player} who moved, the tile they moved from,
+ * and the tile they moved to.</p>
+ *
  * @version 0.1
- * @since 0.1
+ * @author Gilianne Reyes
+ * @param player is the player who moved; never null
+ * @param fromTile is the tile the player moved from.
+ * @param destinationTile is the tile the player moved to.
  */
-public record PlayerMovedEvent(Player player, Tile fromTile, Tile destinationTile) implements GameEvent {}
+public record PlayerMovedEvent(Player player, Tile fromTile, Tile destinationTile)
+      implements GameEvent {}
